Todo.Data.Sorted

Types

data SortedTasks

Sorted tasks.

data SortType

Sort type.

Constructors

SortPriority
Sort by priority.

SortStatus
Sort by status.

SortPriorityStatus
Sort by priority then status.

SortStatusPriority
Sort by status then priority.

Functions

sortTasks

Arguments

:: Maybe SortType
Primary sort type. If Nothing, sorts by status then priority, with all completed tasks appearing at the bottom.

-> RevSort
Reverses the sort.

-> Seq SomeTask
Tasks to sort.

-> SortedTasks

Sorts tasks. Tasks are always sorted by id, so any given sort will take priority, using id as a tie-breaker.

traverseSorted :: (SingleTask -> a) -> (TaskGroup -> a) -> SortedTasks -> Seq a

parseSortType :: MonadFail m => m Text -> m SortType