Todo.Data.TaskStatus

Status

data TaskStatus

Task status. The (commutative) semigroup is based roughly on the ordering

  Completed < NotStarted < InProgress < Blocked

where we take the max. There are exceptionals cases for

  Completed <> NotStarted == InProgress
  Blocked l <> Blocked r == Blocked (l <> r)

The first line is why we do not have a monoid.

Constructors

Completed

NotStarted

InProgress

Blocked (NESet Blocker)
Task is blocked, where the blockers can be another task (represented by TaskId) or a general text description.

parseTaskStatus :: MonadFail m => Text -> m TaskStatus

filterBlockingIds :: NESet Blocker -> Set TaskId

isCompleted :: TaskStatus -> Bool

True iff Completed.

metavar :: IsString a => a

Optics

_Completed :: Prism' TaskStatus ()

_NotStarted :: Prism' TaskStatus ()

_InProgress :: Prism' TaskStatus ()

_Blocked :: Prism' TaskStatus (NESet Blocker)

Blocker

data Blocker

Something that blocks another task.

Constructors

BlockerId TaskId
A task that blocks another task. The task is is the blocker.

BlockerText Text
General description that blocks a task.

parseBlockedTarget :: MonadFail m => Text -> m Blocker

Optics

_BlockerId :: Prism' Blocker TaskId

_BlockerText :: Prism' Blocker Text