Todo.Exception

Types

data BlockedIdRefE

Error for a task t1 referencing task ids that do not exist.

Constructors

MkBlockedIdRefE

Fields

TaskId

t1.id

(NESet TaskId)

t1.refIds

newtype ConfigNotFoundE

Error for config not found.

Constructors

MkConfigNotFoundE OsPath

data DeleteE

Errors when deleting a task.

Constructors

DeleteTaskIdNotFound TaskIdNotFoundE
Attempted to delete an id that was not found

DeleteRefId TaskId (NESet TaskId)
Attempted to delete a task that is referenced by other tasks.

newtype DuplicateIdE

Error for two tasks t1 and t2 having the same id.

Constructors

MkDuplicateIdE TaskId

newtype FoundGroupNotSingleE

Error for finding a group when we wanted a single task.

Constructors

MkFoundGroupNotSingleE TaskId

data IndexNameLookupE

Error when failing to find an index name.

Constructors

MkIndexNameLookupE OsPath Text

newtype TaskIdNotFoundE

Error for not finding a task id in the index.

Constructors

MkTaskIdNotFoundE TaskId

newtype XdgIndexNotFoundE

Error for failing to find the XDG index.

Constructors

MkXdgIndexNotFoundE OsPath

Functions

displayExceptionSkipKnownCS :: Exception e => e -> String

Display exception, skipping callstacks for known exceptions.