Todo.Index

Types

data Index (s :: IndexState)

Todo index.

State

data IndexState

The possible verification states w.r.t internal invariants. These are:

All task ids unique.

All id references exist.

Constructors

IndexStateUnverified
Invariants unverified.

IndexStateVerified
Invariants verified.

type Index𝕌 = Index 'IndexStateUnverified

Unverified index. U+1D54C

type Index𝕍 = Index 'IndexStateVerified

Verified index. U+1D54D

Creation

readIndex :: (HasCallStack, MonadFileReader m, MonadThrow m) => OsPath -> m Index𝕍

Reads the file to an Index.

readTaskList :: (HasCallStack, MonadFileReader m, MonadThrow m) => OsPath -> m (Tuple2 OsPath (Seq SomeTask))

Reads the file to a task list.

fromList :: (HasCallStack, MonadThrow m) => OsPath -> Seq SomeTask -> m Index𝕍

Parses a list into an Index. Throws errors for duplicate ids or id references that do not exist (i.e. Blocked status).

Update

setSomeTaskValue

Arguments

:: forall a (s :: IndexState). Lens' SomeTask a
Lens for the task value we want to set.

-> TaskId
Id for the task to set.

-> a
The new value.

-> Index s
The index.

-> Maybe (Index𝕌, SomeTask)
If successful (task id exists), returns the new index and modified task.

Attempts to set a value for the corresponding task in the index.

setSomeTaskValueMapped

Arguments

:: forall a (s :: IndexState). (Index𝕌 -> Index𝕌)
Index mapper.

-> Lens' SomeTask a
Lens for the task value we want to set.

-> TaskId
Id for the task to set.

-> a
The new value.

-> Index s
The index.

-> Maybe (Tuple2 Index𝕌 SomeTask)
If successful (task id exists), returns the new index and modified task.

Like setSomeTaskValue, except we map the result.

setTaskValue

Arguments

:: forall a (s :: IndexState). Lens' SingleTask a
Lens for the task value we want to set.

-> TaskId
Id for the task to set.

-> a
The new value.

-> Index s
The index.

-> MatchResult Index𝕌 SomeTask
If successful (task id exists), returns the new index and modified task.

Like setSomeTaskValue, except expects a SingleTask. Returns a MatchResult to distinguish between a total failure (task id does not exist at all) vs. a partial failure (task id exists but corresponds to a group), for the purpose of giving a better error message.

Insertion

insert :: forall (s :: IndexState). SomeTask -> Index s -> Index𝕌

Inserts a task into the index at the top-level. Note that this does not check that the id is not a duplicate, hence the need for verification.

insertAtTaskId :: forall (s :: IndexState). GroupTaskId -> SomeTask -> Index s -> Index𝕌

Inserts a new task t1 at the task id. Like insert, this performs no id verification, and the given task id may not exist in the index (in which case t1 will not be added).

Filtering

filterTopLevel :: forall (s :: IndexState). (SomeTask -> Bool) -> Index s -> Index𝕌

Filters the index on some predicate. Note this runs the predicate on the top-level tasks only. Thus if you need to filter on subtasks, the passed predicate will have todo this itself.

filterOnIds

Arguments

:: forall (s :: IndexState). Set TaskId
Ids to take.

-> Index s
Index

-> Index𝕌

Filters the index on the task ids. Included tasks are those with an id in the set, or task groups who have children in the set.

partitionTaskIds

Arguments

:: forall (s :: IndexState). NESet TaskId
TaskIds set s.

-> Index s
Index to partition.

-> Tuple2 Index𝕌 Index𝕌
(Tasks in s, Tasks not in s)

Partitions the Index on the TaskId set. The left result is all tasks whose ids belong in the set and tasks who have a some parent in the set.

The right set is all tasks whose ids do not belong in the set.

partition

Arguments

:: forall (s :: IndexState). (SomeTask -> Bool)
Predicate p.

-> Index s
Index to partition.

-> Tuple2 Index𝕌 Index𝕌
(p == True, p == False)

Partitions the Index according to the predicate p. If p t is true for some t then t is added to the left result, unchanged. This means that, for instance, the entire task group is returned without checking any children.

On the other hand, if p t returns false for some task group t, then we check subtasks. Any subtasks returning true will be added to the left result, whereas t' will be added to the right result, where t' is the original group t with the children that also return false.

Query

Lookup

lookup :: forall (s :: IndexState). TaskId -> Index s -> Maybe SomeTask

Looks up the TaskId in the Index.

data GroupTaskId

Newtype for a group task id.

findGroupTaskId :: forall (s :: IndexState). TaskId -> Index s -> Either Text GroupTaskId

Returns a GroupTaskId for the parameter iff it corresponds to a extant group TaskId. Otherwise returns an error message.

Membership

member :: forall (s :: IndexState). TaskId -> Index s -> Bool

Returns True iff the TaskId exists in the index.

(∈) :: forall (s :: IndexState). TaskId -> Index s -> Bool infix 4

Operator alias for member. U+2216.

notMember :: forall (s :: IndexState). TaskId -> Index s -> Bool

Negation of member.

(∉) :: forall (s :: IndexState). TaskId -> Index s -> Bool infix 4

Negation of (∈). U+2209.

Elimination

writeIndex :: (HasCallStack, MonadFileWriter m) => Index𝕍 -> m ()

Writes the index to the path.

toList :: forall (s :: IndexState). Index s -> (OsPath, Seq SomeTask)

Returns a list representation of the index.

Verification

unverify :: forall (s :: IndexState). Index s -> Index𝕌

Forgets verification status on an Index.

verify :: forall m (s :: IndexState). (HasCallStack, MonadThrow m) => Index s -> m Index𝕍

Verifies that the index satisfies the invariants.

Misc

getBlockingIds :: forall (s :: IndexState). Index s -> Map TaskId (NESet TaskId)

Returns a map of all blocking ids to blockees. For instance, if tasks t1 and t2 are both blocked by t3, then there should be an entry:

t3 -> [t1, t2]