Todo.Utils

Monadic Combinators

whileM :: Monad m => m Bool -> m a -> m (Seq a)

whileApplyM :: Monad m => a -> m Bool -> (a -> m (a, b)) -> m (a, Seq b)

whileApplySetM :: forall m a b. (Monad m, Ord b) => a -> m Bool -> (a -> m (a, b)) -> m (a, Set b)

whileM_ :: Monad m => m Bool -> m a -> m ()

Optics

traversal :: Traversable f => Traversal (f a) (f b) a b

Traversal for any Traversable.

listTraversal :: Traversal (List a) (List b) a b

listPredTraversal :: (a -> Bool) -> Traversal (List a) (List b) a b

seqTraversal :: Traversal (Seq a) (Seq b) a b

seqPredTraversal :: (a -> Bool) -> Traversal (Seq a) (Seq b) a b

neSetTraversal :: forall a b. Ord b => Traversal (NESet a) (NESet b) a b

Traverses an NESet.

Preview + Over

overPreview' :: forall k (is :: IxList) s a. (Is k An_AffineFold, Is k A_Setter) => Optic k is s s a a -> (a -> a) -> s -> Maybe s

Like over', except we run the set' on the result of a preview. This is intended for updates that can fail. For example, we want Index update to fail if we do not find a task with the desired id. The Ixed class, however, provides an AffineTraversal' s a i.e sets always succeed.

We therefore generalize to AffineFold and Setter, and run the set on the preview.

setPreview' :: forall k (is :: IxList) s a. (Is k An_AffineFold, Is k A_Setter) => Optic k is s s a a -> a -> s -> Maybe s

set' with overPreview'.

overPreviewNode'

Arguments

:: AffineTraversal' s a
AffineTraversal from a "root type" s to node a.

-> Lens' a b
Lens from node a to leaf l.

-> (b -> b)
Leaf modifier.

-> s
Root to modify.

-> Maybe (Tuple2 s a)
The new root s' and inner node a', if the update was successful.

Like overPreview', except we also return the updated "inner node", if the update succeeds.

setPreviewNode' :: AffineTraversal' s a -> Lens' a b -> b -> s -> Maybe (Tuple2 s a)

overPreviewNode' that sets the field.

data MatchResult s a

Optics matching result.

Constructors

MatchFailure
Match completed failed.

MatchPartial a
Match "partially" succeeded (context-dependent).

MatchSuccess s a
Match completely succeeded.

overPreviewPartialNode'

Arguments

:: AffineTraversal' s a
AffineTraversal from a "root type" s to node a.

-> AffineTraversal' a b
AffineTraversal from node a to leaf l.

-> (b -> b)
Leaf modifier.

-> s
Root to modify.

-> MatchResult s a
The new root s' and inner node a', if the update was successful.

Like overPreviewNode', except the inner optic is an AffineTraversal, not a Lens. The overall AffineTraversal is split so that we can distinguish between a complete failure (MatchFailure) and a partial failures (MatchParial).

This is used, for instance, to distinguish between:

Failing to find a Task given a TaskId.

Finding a TaskGroup with the TaskId when we wanted a SingleTask.

for the purposes of giving a better error message.

setPreviewPartialNode' :: AffineTraversal' s a -> AffineTraversal' a b -> b -> s -> MatchResult s a

Optics

_MatchSuccess :: Prism (MatchResult s a) (MatchResult t a) (Tuple2 s a) (Tuple2 t a)

_MatchPartial :: Prism' (MatchResult s a) a

_MatchFailure :: Prism' (MatchResult s a) ()

Aeson

validateKeys :: Set Key -> Object -> Parser ()

Validates that the keys in the object are only those we expect i.e. fails if the Object contains unexpected keys.